WebApplication/requirements.txt
altair==4.2.0
arabert==1.0.1
arabic-reshaper==2.1.3
attrs==22.2.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
codetiming==1.3.0
commonmark==0.9.1
decorator==5.1.1
emoji==1.4.2
entrypoints==0.4
farasapy==0.0.14
filelock==3.8.2
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
MarkupSafe==2.1.1
nltk==3.8
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
preprocess==2.0.0
protobuf==3.20.2
PyArabic==0.6.15
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
zipp==3.11.0
